cassandra-utils/requirements.txt
cassandra-driver==3.7.1
config==0.3.9
futures==3.0.5
six==1.10.0
PyYAML==3.12
paramiko==2.1.1